fix for #5 Error in updateSubscriptions Request
I added the Description and Invoice so these two fields should be
updated in the call.

However, the last sentence in the description for this issue is not
correct.

"It also includes some elements that can't be changed such as schedule,
occurences, and amount."

We do allow the occurrence and amount to be updated.  Here are the
fields are NOT allowed to be updated:

BillingInterval
BillingIntervalUnits (month, week etc.)
StartsOn (the day of subscription is billed)
TrialAmount
TrialBillingCycles